+++ backend/src/main/java/hanglog/auth/service/AuthService.java
@@ -14,8 +14,8 @@ import hanglog.auth.domain.oauthuserinfo.OauthUserInfo;
 import hanglog.auth.domain.repository.RefreshTokenRepository;
 import hanglog.global.exception.AuthException;
 import hanglog.member.domain.Member;
+import hanglog.member.domain.MemberDeleteEvent;
 import hanglog.member.domain.repository.MemberRepository;
-import hanglog.member.event.MemberDeleteEvent;
 import hanglog.share.domain.repository.SharedTripRepository;
 import hanglog.trip.domain.repository.CustomTripRepository;
 import hanglog.trip.domain.repository.PublishedTripRepository;
